창 크기 고정 완료
씬 구조 바꾸기 진행 중
더블 버퍼링 구현 예정

+++ Survive38s/Time.h
@@ -2,4 +2,4 @@
 
 void InitTime();
 void UpdateTime();
-ULONGLONG GetDeltaTime();
+double GetDeltaTime();
